Bugfixes related to NSGA-II. - cycle times were 0, now fixed - deviation of ct's was 0, now fixes (added extra variable in sim runner) - pareto front graph now shows a seemingly correct distribution.
- General changes -> conform syntax of prosimos output logs...
- BPMN parser -> dont take resource cost out of XES, rather out of BPMN

+++ support_modules/file_manager.py
@@ -230,7 +230,7 @@ def read_genetic_stats_file(log_name):
                     if block_count:
                         if row[1] not in resource_pools:
                             resource_pools[row[1]] = list()
-                        resource_pools[row[1]].append(ResourceInfo(row[2], int(row[3]), float(row[4]), float(row[5])))
+                        resource_pools[row[1]].append(ResourceInfo(row[2], float(row[3]), float(row[4]), float(row[5])))
                 return [explored_solutions, resource_pools]
     except IOError:
         return None
